Styling Changes
Redid the Hamburger Menu animation. Looks much better now but the center spans need some help with their opacity settings.

Made a good amount of changes to the styling of the new, edit, and show pages.

Moved setImageVal() out of CanvasActual and into a separate js file. This file will be used for general functions, CanvasActual needs to be completely separate.