Styled User Index
Changed styling on user index along with edits to the mediaquery for the user list items.

Bug when inspecting using the mobile tester in chrome. Sometimes layout is correct, sometimes the body width only covers 2/3 of the screen even when set to 100vw !important. Hopefully this is just a chrome inspect quirk.